// app/src/api/model/filterOperatorType.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Economicon API
 * OpenAPI spec version: 0.1.0
 */

export type FilterOperatorType = typeof FilterOperatorType[keyof typeof FilterOperatorType];


export const FilterOperatorType = {
  equals: 'equals',
  notEquals: 'notEquals',
  greaterThan: 'greaterThan',
  lessThan: 'lessThan',
  greaterThanOrEquals: 'greaterThanOrEquals',
  lessThanOrEquals: 'lessThanOrEquals',
} as const;